+++ packages/admin/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Fixed
+
+- When saving a discount for the first time using BuyXGetY, the save method will now ensure the id is set.
+
 ## 0.2.4
 
 ### Changed